Move vars module; write more docs
