--- fidget/src/core/context/var.rs
@@ -1,148 +1,0 @@
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
-
-/// A `Var` represents a value which can vary during evaluation
-///
-/// We pre-define common variables (e.g. `X`, `Y`, `Z`) but also allow for fully
-/// customized values.
-///
-/// Variables are "global", in that every instance of `Var::X` represents the
-/// same thing.  To generate a "local" variable, [`Var::new`] picks a random
-/// 64-bit value, which is very unlikely to collide with anything else.
-#[allow(missing_docs)]
-#[derive(Copy, Clone, Debug, Hash, Eq, PartialEq, Ord, PartialOrd)]
-pub enum Var {
-    X,
-    Y,
-    Z,
-    V(u64),
-}
-
-impl Var {
-    /// Returns a new variable, with a random 64-bit index
-    ///
-    /// The odds of collision with any previous variable are infintesimally
-    /// small; if you are generating billions of random variables, something
-    /// else in the system is likely to break before collisions become an issue.
-    #[allow(clippy::new_without_default)]
-    pub fn new() -> Self {
-        let v: u64 = rand::random();
-        Var::V(v)
-    }
-}
-
-impl std::fmt::Display for Var {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Var::X => write!(f, "X"),
-            Var::Y => write!(f, "Y"),
-            Var::Z => write!(f, "Z"),
-            Var::V(v) if *v < 256 => write!(f, "v_{v}"),
-            Var::V(v) => write!(f, "V({v:x})"),
-        }
-    }
-}
-
-/// Map from [`Var`] to a particular value
-///
-/// This is equivalent to a
-/// [`HashMap<Var, T>`](std::collections::HashMap) and as such does not include
-/// per-function documentation.
-///
-/// The advantage over a `HashMap` is that for common variables (`X`, `Y`, `Z`),
-/// no allocation is required.
-#[derive(Serialize, Deserialize)]
-pub struct VarMap<T> {
-    x: Option<T>,
-    y: Option<T>,
-    z: Option<T>,
-    v: HashMap<u64, T>,
-}
-
-impl<T> Default for VarMap<T> {
-    fn default() -> Self {
-        Self {
-            x: None,
-            y: None,
-            z: None,
-            v: HashMap::default(),
-        }
-    }
-}
-
-#[allow(missing_docs)]
-impl<T> VarMap<T> {
-    pub fn new() -> Self {
-        Self::default()
-    }
-    pub fn len(&self) -> usize {
-        self.x.is_some() as usize
-            + self.y.is_some() as usize
-            + self.z.is_some() as usize
-            + self.v.len()
-    }
-    pub fn is_empty(&self) -> bool {
-        self.x.is_none()
-            && self.y.is_none()
-            && self.z.is_none()
-            && self.v.is_empty()
-    }
-    pub fn get(&self, v: &Var) -> Option<&T> {
-        match v {
-            Var::X => self.x.as_ref(),
-            Var::Y => self.y.as_ref(),
-            Var::Z => self.z.as_ref(),
-            Var::V(v) => self.v.get(v),
-        }
-    }
-
-    pub fn get_mut(&mut self, v: &Var) -> Option<&mut T> {
-        match v {
-            Var::X => self.x.as_mut(),
-            Var::Y => self.y.as_mut(),
-            Var::Z => self.z.as_mut(),
-            Var::V(v) => self.v.get_mut(v),
-        }
-    }
-
-    pub fn entry(&mut self, v: Var) -> VarMapEntry<T> {
-        match v {
-            Var::X => VarMapEntry::Option(&mut self.x),
-            Var::Y => VarMapEntry::Option(&mut self.y),
-            Var::Z => VarMapEntry::Option(&mut self.z),
-            Var::V(v) => VarMapEntry::Hash(self.v.entry(v)),
-        }
-    }
-}
-
-pub enum VarMapEntry<'a, T> {
-    Option(&'a mut Option<T>),
-    Hash(std::collections::hash_map::Entry<'a, u64, T>),
-}
-
-impl<'a, T> VarMapEntry<'a, T> {
-    pub fn or_insert(self, default: T) -> &'a mut T {
-        match self {
-            VarMapEntry::Option(o) => match o {
-                Some(v) => v,
-                None => {
-                    *o = Some(default);
-                    o.as_mut().unwrap()
-                }
-            },
-            VarMapEntry::Hash(e) => e.or_insert(default),
-        }
-    }
-}
-
-impl<T> std::ops::Index<&Var> for VarMap<T> {
-    type Output = T;
-    fn index(&self, v: &Var) -> &Self::Output {
-        match v {
-            Var::X => self.x.as_ref().unwrap(),
-            Var::Y => self.y.as_ref().unwrap(),
-            Var::Z => self.z.as_ref().unwrap(),
-            Var::V(v) => &self.v[v],
-        }
-    }
-}
